Update to 0.3
- Add quests
- Rename game player inventory to mercury player inventory
- Add quest placeholders
- Add cache to database
- Add sidebar
- Add player and quest tick tasks

+++ src/main/java/me/adamix/mercury/server/player/stats/Statistics.java
@@ -46,9 +46,9 @@ public class Statistics {
 
 		map.forEach((categoryString, valueObject) -> {
 			StatisticCategory category = StatisticCategory.valueOf(categoryString);
-			Map<String, Float> valueMap = (Map<String, Float>) valueObject;
+			Map<String, Double> valueMap = (Map<String, Double>) valueObject;
 			valueMap.forEach((name, value) -> {
-				statistics.set(category, name, value);
+				statistics.set(category, name, value.floatValue());
 			});
 		});
